refactor: merge CleanUpLandedFlights into CleanUpFlights
Combines landed flight cleanup with lost-flight removal (moved from
ProcessFlightsHandler) and adds stale FlightDataRecord cleanup for
unactivated flights. Removal is now driven entirely by LastSeen timeout
rather than flight plan events.
