+++ source/Maestro.Core.Tests/Handlers/ProcessFlightHandlerTests.cs
@@ -459,90 +459,6 @@ public class ProcessFlightHandlerTests(ClockFixture clockFixture)
     }
 
     [Fact]
-    public async Task WhenAFlightFdrDataIsStale_ItIsRemovedFromSequence()
-    {
-        // Arrange - #98: disconnected aircraft should be removed after timeout
-        var airportConfiguration = GetDefaultAirportConfiguration(lostFlightTimeoutMinutes: 10);
-        var clock = clockFixture.Instance;
-        var flight = new FlightBuilder("QFA123")
-            .WithState(State.Frozen)
-            .WithFeederFix("RIVET")
-            .WithFeederFixEstimate(clock.UtcNow().AddMinutes(5))
-            .Build();
-
-        var (sessionManager, session, sequence) = new SessionBuilder(airportConfiguration)
-            .WithSequence(s => s.WithFlight(flight))
-            .Build();
-
-        session.FlightDataRecords["QFA123"] = MakeRecord("QFA123",
-            lastSeen: clock.UtcNow().AddMinutes(-11));
-
-        var handler = GetHandler(airportConfiguration, sessionManager, clock);
-
-        // Act
-        await handler.Handle(new ProcessFlightsRequest("YSSY"), CancellationToken.None);
-
-        // Assert
-        sequence.Flights.ShouldBeEmpty("flight not seen within lost timeout should be removed");
-    }
-
-    [Fact]
-    public async Task WhenAManuallyInsertedFlightHasNoFdrData_ItIsNotRemoved()
-    {
-        // Arrange
-        var airportConfiguration = GetDefaultAirportConfiguration(lostFlightTimeoutMinutes: 10);
-        var clock = clockFixture.Instance;
-        var flight = new FlightBuilder("****01*")
-            .AsManuallyInserted()
-            .WithState(State.Frozen)
-            .WithTargetLandingTime(clock.UtcNow().AddMinutes(10))
-            .Build();
-
-        var (sessionManager, _, sequence) = new SessionBuilder(airportConfiguration)
-            .WithSequence(s => s.WithFlight(flight))
-            .Build();
-
-        // No FlightDataRecord for dummy flight
-
-        var handler = GetHandler(airportConfiguration, sessionManager, clock);
-
-        // Act
-        await handler.Handle(new ProcessFlightsRequest("YSSY"), CancellationToken.None);
-
-        // Assert
-        sequence.Flights.ShouldHaveSingleItem("manually inserted flights should never be removed by ProcessFlightsHandler");
-    }
-
-    [Fact]
-    public async Task WhenAFlightFdrDataIsStale_ButFlightHasLanded_ItIsNotRemovedByProcessFlights()
-    {
-        // Arrange - landed flights are handled by CleanUpLandedFlightsRequestHandler
-        var airportConfiguration = GetDefaultAirportConfiguration(lostFlightTimeoutMinutes: 10);
-        var clock = clockFixture.Instance;
-        var flight = new FlightBuilder("QFA123")
-            .WithState(State.Landed)
-            .WithFeederFix("RIVET")
-            .WithFeederFixEstimate(clock.UtcNow().AddMinutes(-20))
-            .WithLandingTime(clock.UtcNow().AddMinutes(-2))
-            .Build();
-
-        var (sessionManager, session, sequence) = new SessionBuilder(airportConfiguration)
-            .WithSequence(s => s.WithFlight(flight))
-            .Build();
-
-        session.FlightDataRecords["QFA123"] = MakeRecord("QFA123",
-            lastSeen: clock.UtcNow().AddMinutes(-15));
-
-        var handler = GetHandler(airportConfiguration, sessionManager, clock);
-
-        // Act
-        await handler.Handle(new ProcessFlightsRequest("YSSY"), CancellationToken.None);
-
-        // Assert
-        sequence.Flights.ShouldHaveSingleItem("landed flights should be left for CleanUpLandedFlightsRequestHandler");
-    }
-
-    [Fact]
     public async Task WhenAFlightHasNoFdrData_StateIsStillUpdatedBasedOnTime()
     {
         // Arrange - #84: state transitions should happen on timer, not FDR arrival
