+++ source/Maestro.Core/Sessions/Contracts/CleanUpFlightsRequest.cs
@@ -1,0 +1,5 @@
+using MediatR;
+
+namespace Maestro.Core.Sessions.Contracts;
+
+public record CleanUpFlightsRequest(string AirportIdentifier) : IRequest;
