--- source/Maestro.Core/Sessions/Contracts/CleanUpLandedFlightsRequest.cs
@@ -1,5 +1,0 @@
-using MediatR;
-
-namespace Maestro.Core.Sessions.Contracts;
-
-public record CleanUpLandedFlightsRequest(string AirportIdentifier) : IRequest;
