+++ source/Maestro.Core/Sessions/Session.cs
@@ -82,7 +82,7 @@ public class Session : IAsyncDisposable
             {
                 await _mediator.Send(new TrySwapRunwayModesRequest(AirportIdentifier), cancellationToken);
                 await _mediator.Send(new ProcessFlightsRequest(AirportIdentifier), cancellationToken);
-                await _mediator.Send(new CleanUpLandedFlightsRequest(AirportIdentifier), cancellationToken);
+                await _mediator.Send(new CleanUpFlightsRequest(AirportIdentifier), cancellationToken);
 
                 await Task.Delay(TimeSpan.FromSeconds(30), cancellationToken);
             }
